requirements-dev.txt
# base
-r requirements.txt

# dev
black==19.3b0
django-extensions==2.2.5
flake8-isort==2.7.0
flake8-quotes==2.0.1
flake8==3.7.7
ipdb==0.12.2
isort==4.3.21
pytest==5.2.2
pytest-django==3.6.0
